Reorganized some files.
