+++ src/config/sceneSetup.js
